refactor(runsheet): extract shared link renderer from PCO viewer
Move the PCO run sheet's URL detection, SelectableText, Dropbox video card,
and rich link-preview logic into a shared util (runSheetLinks.tsx) so the
native viewer can reuse the exact same treatment instead of duplicating it.
Pure refactor of RunSheetScreen — no behavior change.
